TrackedPoint & subclasses (Pump, LevelMeter)
